% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_de.R
\name{fit_de}
\alias{fit_de}
\title{fit_de}
\usage{
fit_de(objective, start = NULL, lower = NULL, upper = NULL,
  hessian = FALSE, sigma = 0.1, n_pop = 50L, opt_args = list(),
  obj_args = list(), ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{lower}{numeric vector; lower bounds on parameters}

\item{upper}{numeric vector; upper bounds on parameters}

\item{hessian}{logical; if TRUE, hessian is calculated at solution. Default = FALSE}

\item{sigma}{numeric vector; standard deviation to create random initial population}

\item{n_pop}{numeric vector; size of population}

\item{opt_args}{list; list of arguments passed to DEoptim, see RcppDE::DEoptim for details}

\item{obj_args}{list; list of arguments to pass on to objective function}

\item{...}{further arguments passed to objective}
}
\description{
runs optimization using RcppDE package (Rcpp version of DEoptim)
}
