% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ga.R
\name{fit_ga}
\alias{fit_ga}
\title{fit_ga}
\usage{
fit_ga(objective, start = NULL, lower = NULL, upper = NULL,
  hessian = T, sigma = 0.1, n_pop = 50L, minimize = T,
  opt_args = list(), obj_args = list(), ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{lower}{numeric vector; lower bounds on parameters}

\item{upper}{numeric vector; upper bounds on parameters}

\item{hessian}{logical; if TRUE, find the hessian at the optimum}

\item{sigma}{numeric; standard deviation to create random initial population}

\item{n_pop}{integer; population size}

\item{minimize}{logical; if TRUE, maximize -1 * objective}

\item{opt_args}{list; list of arguments to pass to ga, see GA::ga for details}

\item{obj_args}{list; list of arguments to pass on to objective function}

\item{...}{further arguments passed to objective}
}
\description{
runs optimization using GA package
}
