% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gensa.R
\name{fit_gensa}
\alias{fit_gensa}
\title{fit_gensa}
\usage{
fit_gensa(objective, start, lower, upper, control = list(),
  hessian = T, ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{lower}{numeric vector; lower bounds on parameters}

\item{upper}{numeric vector; upper bounds on parameters}

\item{control}{list; list of control parameters, see GenSA function for details}

\item{...}{further arguments passed to objective}
}
\description{
runs optimization using GenSA package
}
