% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_opt.R
\name{fit_opt}
\alias{fit_opt}
\title{fit_opt}
\usage{
fit_opt(
  objective,
  start,
  lower = NULL,
  upper = NULL,
  hessian = FALSE,
  package = "optimx",
  method = NULL,
  restart = F,
  consecutive = 0L,
  max_runs = Inf,
  sigma = 0.1,
  tol = 1e-05,
  verbose = T,
  opt_args = list(),
  obj_args = list(),
  ...
)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{lower}{numeric vector; lower bounds for parameters}

\item{upper}{numeric vector; upper bounds for parameters}

\item{hessian}{logical; if TRUE, calculate hessian at solution}

\item{package}{string; package to use, see Details. Default = "optimx"}

\item{method}{string; method to use from package}

\item{restart}{logical; if TRUE, restart optimization until no more improvement. Default = FALSE}

\item{max_runs}{integer; maximum number of optimx runs. Default = Inf}

\item{sigma}{numeric; standard deviation for random perturbation of starting values}

\item{verbose}{logical; if TRUE, print result summary after each iteration}

\item{opt_args}{list; list of arguments to be passed to optimx}

\item{obj_args}{list; list of arguments to be passed to objective function}

\item{...}{further arguments objective}

\item{conescutive}{integer; stop after this number of consecutive runs without improvement. Default = 0}

\item{tolerance}{numeric; new fit must change objective by 'tolerance' to be considered an improvement}
}
\description{
nonlinear optimization: wrapper for optimx and nloptr
}
\details{
Package options = "optimx", "nloptr"
}
