% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_optimx.R
\name{fit_optimx}
\alias{fit_optimx}
\title{fit_optimx}
\usage{
fit_optimx(objective, start, max_runs = 5L, consecutive = 0L,
  sigma = 0.1, tol = 1e-05, verbose = T, ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{max_runs}{integer; maximum number of optimx runs}

\item{sigma}{numeric; standard deviation for random perturbation of starting values}

\item{verbose}{logical; if TRUE, print result summary after each iteration}

\item{...}{further arguments passed to run_optimx, optimx, and objective}

\item{conescutive}{integer; stop after this number of consecutive runs without improvement}

\item{tolerance}{numeric; new fit must change objective by 'tolerance' to be considered an improvement}
}
\description{
runs optimization using optimx package
}
