% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{opt}
\alias{opt}
\title{opt}
\usage{
opt(objective, start, package = "optimx", failcode = NULL,
  max_runs = 5L, consecutive = 0L, sigma = 0.1, tol = 1e-05,
  verbose = T, ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{package}{string; package to use, see Details. Default = "optimx"}

\item{failcode}{integer vector; convergence code from optimizer to be considered failure}

\item{max_runs}{integer; maximum number of optimx runs}

\item{sigma}{numeric; standard deviation for random perturbation of starting values}

\item{verbose}{logical; if TRUE, print result summary after each iteration}

\item{...}{further arguments passed to run_optimx, optimx, and objective}

\item{conescutive}{integer; stop after this number of consecutive runs without improvement}

\item{tolerance}{numeric; new fit must change objective by 'tolerance' to be considered an improvement}
}
\description{
nonlinear optimization: wrapper for optimx and nloptr
}
\details{
Package options = "optimx", "nloptr"
}
