% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nloptr.R
\name{run_nloptr}
\alias{run_nloptr}
\title{run_nloptr}
\usage{
run_nloptr(objective, start, lower = NULL, upper = NULL,
  algorithm = "neldermead", opts = list(), hessian = T, ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{algorithm}{string; method to use. See details. Default = "neldermead"}

\item{opts}{list; list of options. See nloptr::nloptr.print.options() for details.}

\item{hessian}{logical; if TRUE, find the hessian at the optimum. Default = T}

\item{...}{further arguments passed to nloptr method and objective}
}
\description{
runs optimization using nloptr package
}
\details{
Algorithm options =
"bobyqa",
"ccsaq",
"cobyla",
"crs2lm",
"direct",
"isres",
"lbfgs",
"mma",
"neldermead",
"newuoa",
"sbplx",
"slsqp",
"stogo",
"tnewton",
"varmetric"
}
