% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optim.R
\name{run_optim}
\alias{run_optim}
\title{run_optim}
\usage{
run_optim(
  objective,
  start,
  lower = -Inf,
  upper = Inf,
  hessian = FALSE,
  method = "Nelder-Mead",
  control = list(),
  ...
)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{lower}{numeric vector; lower bounds for parameters}

\item{upper}{numeric vector; upper bounds for parameters}

\item{hessian}{logical; if TRUE, find the hessian at the optimum}

\item{method}{string; method to use. See optim for details}

\item{control}{list; a list of control parameters, see optim for details}

\item{...}{further arguments passed to optim and objective}
}
\description{
runs optimization using optim package
}
