% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimx.R
\name{run_optimx}
\alias{run_optimx}
\title{run_optimx}
\usage{
run_optimx(objective, start, method = "nmkb", hessian = T,
  control = list(), failcode = 9999, ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{method}{string; method to use. See optimx for details}

\item{hessian}{logical; if TRUE, find the hessian at the optimum}

\item{control}{list; a list of control parameters, see optimx for details}

\item{...}{further arguments passed to optimx and objective}
}
\description{
runs optimization using optimx package
}
