% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pracma.R
\name{run_pracma}
\alias{run_pracma}
\title{run_pracma}
\usage{
run_pracma(objective, start, lower = NULL, upper = NULL,
  hessian = FALSE, method = "fminsearch", ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{hessian}{logical; if TRUE, find the hessian at the optimum}

\item{method}{string; method to use. See details. Default = "fminsearch".}

\item{...}{further arguments passed to optimx and objective}
}
\description{
runs optimization using optimx package
}
\details{
method options = "fminsearch", "fminunc", "fmincon", "fminbnd"
}
