% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gensa.R
\name{fit_gensa}
\alias{fit_gensa}
\title{fit_gensa}
\usage{
fit_gensa(objective, start, lower, upper, hessian = FALSE,
  opt_args = list(), obj_args = list(), ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; starting parameters}

\item{lower}{numeric vector; lower bounds on parameters}

\item{upper}{numeric vector; upper bounds on parameters}

\item{hessian}{logical; if TRUE, calculate hessian at solution. Default = FALSE}

\item{opt_args}{list; list of arguments passed on to GenSA, see GenSA function for details}

\item{obj_args}{list; list of arguments to pass on to objective function}

\item{...}{further arguments passed to objective}
}
\description{
runs optimization using GenSA package
}
