% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(objective, start = NULL, method = "Nelder-Mead", aic = T,
  bic = F, n_obs = NULL, as_data_frame = F, ...)
}
\arguments{
\item{objective}{function; the objective function to minimize}

\item{start}{numeric vector; parameter vector}

\item{method}{string; fitting method; by default, Nelder-Mead; see Details}

\item{aic}{logical; if TRUE, calculate Akaike Information Criterion}

\item{bic}{logical; if TRUE, calculate Bayes Information Criterion}

\item{n_obs}{integer; the number of observations (used to calculate BIC)}

\item{...}{further arguments passed to fit_optimx, run_optimx, optimx, and objective}
}
\description{
run optimization algorithm on given objective function
}
